fix: move structs to types folder
